IN work inprogress

+++ src/ee/executors/tuplescanexecutor.cpp
@@ -75,7 +75,7 @@ bool TupleScanExecutor::p_execute(const NValueArray &params) {
     TableTuple temp_tuple = output_temp_table->tempTuple();
     const std::vector<int>& paramIdxs = node->getParamIdxs();
     assert(paramIdxs.size() == output_temp_table->schema()->columnCount());
-    for (size_t i = 0; i < paramIdxs.size(); ++i)
+    for (int i = 0; i < paramIdxs.size(); ++i)
     {
         temp_tuple.setNValue(i, params[paramIdxs[i]]);
     }
